leiloes_rmq_api/pyproject.toml
[tool.poetry]
name = "leiloes_rmq"
version = "0.1.0"
description = "Microservices auction system using RabbitMQ"
authors = ["Eduardo Soviersovski",  "Pedro Tortola"]
readme = "README.md"
packages = [{ include = "*" }]

[tool.poetry.dependencies]
python = "^3.12"
pika = "==1.3.2"
requests = "2.31.0"
pycryptodome = "^3.23.0"
cryptography = "^45.0.6"

[build-system]
requires = ["poetry-core>=1.4.0"]
build-backend = "poetry.core.masonry.api"
